### 2016-12-05 #129 @whedon
[PRE REVIEW]: pyuvdata: an interface for astronomical interferometeric datasets in python
**Submitting author:** bhazelton (<a href="http://orcid.org/0000-0001-7532-645X">Bryna J. Hazelton</a>)
**Repository:** <a href="https://github.com/HERA-Team/pyuvdata" target ="_blank">https://github.com/HERA-Team/pyuvdata</a>
**Version:** v1.0

Thanks for submitting your paper to JOSS bhazelton. The JOSS editor (shown at the top of this issue) will work with you on this issue to find a reviewer for your submission before creating the main review issue.
