### 2016-12-06 348c99a
first draft of cal.py module
