### 2016-12-06 8c3725f
start test_cal.py for testing UVCal
