### 2016-12-06 72361c7
update imports and remove bugs in test_cal.py
