### 2016-12-06 0c2229c
update imports in cal.py and super call.
