### 2016-12-07 423149a
conform to pep8 in cal.py
