+++ uvdata/cal.py
@@ -1,7 +1,7 @@
-#comment notes
-#1. Currently supporting direction independent gains. Future versions may
+# comment notes
+# 1. Currently supporting direction independent gains. Future versions may
 #   support direction dependent gains with list of such objects.
-#2.
+# 2.
 import numpy as np
 from uvbase import UVBase
 import parameter as uvp
@@ -10,13 +10,17 @@ import parameter as uvp
 class UVCal(UVBase):
     """ A class defining a calibration """
     def __init__(self):
-        self._Nfreqs = uvp.UVParameter('Nfreqs', description='Number of frequency channels',
+        self._Nfreqs = uvp.UVParameter('Nfreqs',
+                                       description='Number of frequency channels',
                                        expected_type=int)
-        self._Npols = uvp.UVParameter('Npols', description='Number of polarizations',
+        self._Npols = uvp.UVParameter('Npols',
+                                      description='Number of polarizations',
                                       expected_type=int)
-        self._Ntimes = uvp.UVParameter('Ntimes', description='Number of times',
+        self._Ntimes = uvp.UVParameter('Ntimes',
+                                       description='Number of times',
                                        expected_type=int)
-        self._history = uvp.UVParameter('history', description='String of history, units English',
+        self._history = uvp.UVParameter('history',
+                                        description='String of history, units English',
                                         form='str', expected_type=str)
 
         desc = ('Number of antennas with data present (i.e. number of unique '
@@ -30,31 +34,32 @@ class UVCal(UVBase):
                 'to ant_1_array and ant_2_array). There must be one entry '
                 'here for each unique entry in ant_1_array and '
                 'ant_2_array, but there may be extras as well.')
-        self._antenna_names = uvp.UVParameter('antenna_names', description=desc,
+        self._antenna_names = uvp.UVParameter('antenna_names',
+                                              description=desc,
                                               form=('Nants_telescope',),
                                               expected_type=str)
 
-        desc = ('List of integer antenna numbers corresponding to antenna_names, '
+        desc = ('List of integer antenna numbers corresponding to'
+                'antenna_names,'
                 'shape (Nants_telescope). There must be one '
                 'entry here for each unique entry in ant_1_array and '
                 'ant_2_array, but there may be extras as well.')
-        self._antenna_numbers = uvp.UVParameter('antenna_numbers', description=desc,
+        self._antenna_numbers = uvp.UVParameter('antenna_numbers',
+                                                description=desc,
                                                 form=('Nants_telescope',),
                                                 expected_type=int)
 
         desc = ('Number of antennas in the array. May be larger ' +
                 'than the number of antennas with data')
         self._Nants_telescope = uvp.UVParameter('Nants_telescope',
-                                                description=desc, expected_type=int)
-
-
+                                                description=desc,
+                                                expected_type=int)
 
         desc = 'Array of frequencies, shape (Nspws, Nfreqs), units Hz'
         self._freq_array = uvp.UVParameter('freq_array', description=desc,
                                            form=('Nspws', 'Nfreqs'),
                                            expected_type=np.float,
                                            tols=1e-3)  # mHz
-
 
         desc = ('Array of polarization integers, shape (Npols). '
                 'AIPS Memo 117 says: stokes 1:4 (I,Q,U,V);  '
@@ -75,12 +80,12 @@ class UVCal(UVBase):
         desc = ('The convention for applying he calibration solutions to data.'
                 'Divide/multiply uncalibrated data by gains.')
         self._gain_convention = uvp.UVParameter('gain_convention', form='str',
-                                           expected_type=str,
-                                           description=desc,
-                                           sane_vals=['divide', 'multiply'])
+                                                expected_type=str,
+                                                description=desc,
[…]
