### 2016-12-07 93763c2
pep8 fixes
