### 2016-12-08 611ec1f
Add switching methods for cal_type related parameters
