### 2016-12-09 1c2160e
add Nspws parameter into UVCal parameter list.
