### 2016-12-09 903b2c8
Add check method, improve some documentation
