### 2016-12-09 83a147b
add _convert_to_filetype function in cal.py
