### 2016-12-09 5ec4bc1
add in new test class for testing basic parameters of UVCal
