### 2016-12-09 557a255
add in calfits.py which defines class for writing calibrations to FITS files.
CALFITS class subcalsses UVCal.
