### 2016-12-09 ca43608
Put warning print cleanup in uvbase rather than uvdata
