### 2016-12-09 617d159
Move basic tests from test_uvdata to test_uvbase
also move one to test_parameter
