### 2016-12-10 070ebc6
pep8 fixes to calfits.py
