### 2016-12-14 255c339
put comment into issue
