### 2016-12-14 #129 @astrofrog
@arfon - very sorry for the delay, I'm happy to review this. I should have it done by the end of the week.
