### 2016-12-15 #140 @whedon
[REVIEW]: pyuvdata: an interface for astronomical interferometeric datasets in python
**Submitting author:** @bhazelton (<a href="http://orcid.org/0000-0001-7532-645X">Bryna J. Hazelton</a>)
**Repository:** <a href="https://github.com/HERA-Team/pyuvdata" target ="_blank">https://github.com/HERA-Team/pyuvdata</a>
**Version:** v1.0

- [x] **Authorship:** Has the submitting author (bhazelton) made major contributions to the software?
