### 2017-01-07 5a1953e
add in primary header variables to write and column tables.
