### 2017-01-08 #140 @astrofrog
@arfon - sorry for the delay, will review this now!
