### 2017-01-08 HERA-Team/pyuvdata#117 @astrofrog
Make package pip-installable
For ease of installation, it would be good to upload the release to PyPI so that users can simply do:

```
pip install pyuvdata
```

The link to the release in the README is also broken, it should be https://github.com/HERA-Team/pyuvdata/releases/tag/v1.0 - (but this will be moot if you upload the package to PyPI)

**Note:** I am reviewing the JOSS submission (https://github.com/openjournals/joss-reviews/issues/140) and I consider this to be a blocker
