### 2017-01-08 HERA-Team/pyuvdata#119 @astrofrog
Expand tutorial in documentation
At the moment, the only usage example provided in the docs is pretty minimal. This should be expanded and should for example show how to access the data programmatically once it is read in.

The tutorial as-is is not functional, because the import should be:

```python
from uvdata import UVData
```

In addition, I think ``read_miriad`` shouldn't take the ``miriad`` argument? (same comment for writing)

**Note:** I am reviewing the JOSS submission (https://github.com/openjournals/joss-reviews/issues/140) and I consider this to be a blocker.
