### 2017-01-08 HERA-Team/pyuvdata#120 @astrofrog
UVData class missing from API docs
The UVData class is missing from the API docs:

https://pyuvdata.readthedocs.io/en/latest/developer_docs.html

Since this appears to be the core class for the package, it should be documented.

There are other classes (e.g. UVParameter) that aren't documented either.

**Note:** I am reviewing the JOSS submission (https://github.com/openjournals/joss-reviews/issues/140) and I consider this to be a blocker.
