### 2017-01-08 HERA-Team/pyuvdata#122 @astrofrog
Add community guidelines
JOSS requires community guidelines to be present:

*Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support*

This can take the form on a sentence in the README that says that pull requests are welcome, and that people can open issues in the issue tracker (with a link to it).

**Note:** I am reviewing the JOSS submission (https://github.com/openjournals/joss-reviews/issues/140) and I consider this to be a blocker.
