### 2017-01-08 #140 @astrofrog
I've opened a few issues in the target repository and have linked to here for the critical ones (there are four, above). Once these are addressed, I think all the requirements will be met.
