### 2017-01-08 #140 @arfon
> I've opened a few issues in the target repository and have linked to here for the critical ones (there are four, above). Once these are addressed, I think all the requirements will be met.

👍  thanks. @bhazelton - let us know when you've been able to address @astrofrog's review feedback.
