+++ uvdata/ms.py
@@ -1,0 +1,160 @@
+"""Class for reading and writing casa measurement sets."""
+from astropy import constants as const
+from astropy.time import Time
+import numpy as np
+import warnings
+from uvdata import UVData
+import parameter as uvp
+import casacore.tables as tables
+import telescopes
+
+polDict={1:1,2:2,3:3,4:4,5:-1,6:-2,7:-3,8:-4,9:-5,10:-6,11:-7,12:-8}
+#convert from casa stokes integers to pyuvdata
+class MS(UVData):
+    """
+    Defines a class for reading and writing casa measurement sets.
+    Attributs:
+      ms_required_extra: Names of optional MSParameters that are required for casa ms
+    """
+    def _ms_hist_to_string(history_table):
+        '''
+        converts a CASA history table into a string that can be stored as the uvdata history parameter.
+        Returns this string
+        Args: history_table, a casa table object
+        Returns: casa history table converted to a string with \n denoting new lines and \t denoting column breaks
+        '''
+        history_str='APP_PARAMS\tCLI_COMMAND\tAPPLICATION\tMESSAGE\tOBJECT_ID\tOBSERVATION_ID\tORIGIN\tPRIORITY\tTIME\n'
+        app_params=history_table.getcol('APP_PARAMS')['array']
+        cli_command=history_table.getcol('CLI_COMMAND')['array']
+        application=history_table.getcol('APPLICATION')
+        message=history_table.getcol('MESSAGE')
+        obj_id=history_table.getcol('OBJECT_ID')
+        obs_id=history_table.getcol('OBSERVATION_ID')
+        origin=history_table.getcol('ORIGIN')
+        priority=history_table.getcol('PRIORITY')
+        times=history_table.getcol('TIME')
+        #Now loop through columns and generate history string
+        for tbrow in range(len(times)):
+            newline=str(app_params[tbrow]) \
+            +'\t'+str(cli_command[tbrow]) \
+            +'\t'+str(application[tbrow]) \
+            +'\t'+str(message[tbrow]) \
+            +'\t'+str(obj_id[tbrow]) \
+            +'\t'+str(obs_id[tbrow]) \
+            +'\t'+str(origin[tbrow]) \
+            +'\t'+str(priority[tbrow]) \
+            +'\t'+str(times[tbrow])+'\n'
+            history_str+=newline
+        return history_str
+
+
+
+
+    def write_ms():
+        '''
+        writing ms is not yet supported
+        '''
+
+
+
+    def read_ms(self,filename,run_check=True,run_sanity_check=True,data_column='DATA'):
+    '''
+    read in a casa measurement set
+    ARGS:
+        filename: name of the measurement set folder
+        run_check:
+    '''
+            if not os.path.exists(filepath):
+                raise(IOError, filepath + ' not found')
+            #set visibility units
+            if(data_column=='DATA'):
+                self.vis_units.value="uncalib"
+            elif(data_column=='CORRECTED_DATA'):
+                self.vis_units.value="Jy"
+            elif(data_column=='MODEL'):
+                self.vis_units.value="Jy"
+            self.data_column.value=data_column
+            #get spectral window information
+            tb_spws=casacore.tables.table(filename+'/SPECTRAL_WINDOW')
+            freqs=tb_spws.getcol('CHAN_FREQ')
+            self.freq_array.value=freqs
+            self.Nfreqs.value=int(freqs.shape[1])
+            self.channel_width.value=tb_spws.getcol('CHAN_WIDTH')[0,0]
+            self.Nspws.value=int(freqs.shape[0])
+            self.spw_array.value=n.arange(self.Nspws.value)
+            tb_spws.close()
+            #now get the data
+            tb=casacore.tables.table(filename)
[…]
