### 2017-01-10 84fde59
added first draft of read only ms.py. Modified uvdata.py to include an optional  'data_column' uvparam that specifies whether the DATA, MODEL, or CORRECTED_DATA measurement set columns were read into data_array
