### 2017-01-10 649162b
converted uvfits in data folder to test measurement sets. Converted using CASA 4.7.0-1 for mac osx 10.11 importuvfits method
