+++ uvdata/tests/test_ms.py
@@ -1,0 +1,6 @@
+"""Tests for MS object."""
+import nose.tools as nt
+import os
+from uvdata import UVData
+import uvdata.tests as uvtest
+from uvdata.data import DATA_PATH
