### 2017-01-10 5430c6a
update CALFITS.read/write_fits()
- no more spaces in column names.
- pep8 fixes.
- update read_fits function to read fits file.
