### 2017-01-11 d51f12f
add switch in CALFITS.read_uvfits for setting gains vs delays.
