### 2017-01-11 a1d663a
Remove pyfits from dependency list and docs (issue #116)
It is a dependency of aipy, not pyuvdata and is installed as needed when aipy is installed.
