### 2017-01-11 4ef1a09
Remove pyfits from dependency list and docs (issue #116)
It is a dependency of aipy, not pyuvdata and is installed as needed when aipy is installed.
