### 2017-01-11 8d7b893
Fix dependency list formatting on RTD (issue #118)
also fix latest release link in readme
