+++ docs/index.rst
@@ -118,6 +118,7 @@ Install pyuvdata
 For simple installation, the latest stable version is available via pip
 using ``pip install pyuvdata``
 
+
 Optionally install the development version
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
