### 2017-01-11 4e87e93
Fix dependency list formatting on RTD (issue #118)
also fix latest release link in readme
