### 2017-01-11 7bb5c17
Fix dependency list formatting on RTD (issue #118)
also fix latest release link in readme
