### 2017-01-11 0044013
bug fix: make integration_time in read_uvfits be in seconds
