### 2017-01-11 3b27aa6
bug fix: make integration_time in read_uvfits be in seconds
