+++ pyuvdata/uvfits.py
@@ -157,6 +157,7 @@ class UVFITS(UVData):
         except:
             if self.Ntimes > 1:
                 self.integration_time = \
+
                     float(np.diff(np.sort(list(set(self.time_array))))[0]) * 86400
 
 
