### 2017-01-11 2694ceb
fixed bugs in ms.py. Correction to bug in uvfits.py in which interval was read as a fractional julian day (this has also been fixed in master and I will rebase).
