+++ pyuvdata/uvfits.py
@@ -158,6 +158,7 @@ class UVFITS(UVData):
             if self.Ntimes > 1:
                 self.integration_time = \
                     float(np.diff(np.sort(list(set(self.time_array))))[0]) * 86400
+
             else:
                 raise ValueError('integration time not specified and only '
                                  'one time present')
