+++ uvdata/ms.py
@@ -1,14 +1,16 @@
 """Class for reading and writing casa measurement sets."""
 from astropy import constants as const
-from astropy.time import Time
+import astropy.time as time
 import numpy as np
+import os
 import warnings
 from uvdata import UVData
 import parameter as uvp
 import casacore.tables as tables
 import telescopes
 
-polDict={1:1,2:2,3:3,4:4,5:-1,6:-2,7:-3,8:-4,9:-5,10:-6,11:-7,12:-8}
+polDict={1:1,2:2,3:3,4:4,5:-1,6:-3,7:-4,8:-2,9:-5,10:-7,11:-8,12:-6}
+
 #convert from casa stokes integers to pyuvdata
 class MS(UVData):
     """
@@ -16,14 +18,15 @@ class MS(UVData):
     Attributs:
       ms_required_extra: Names of optional MSParameters that are required for casa ms
     """
-    def _ms_hist_to_string(history_table):
+    ms_required_extra=['data_column','antenna_positions']
+    def _ms_hist_to_string(self,history_table):
         '''
         converts a CASA history table into a string that can be stored as the uvdata history parameter.
         Returns this string
         Args: history_table, a casa table object
-        Returns: casa history table converted to a string with \n denoting new lines and \t denoting column breaks
+        Returns: casa history table converted to a string with \n denoting new lines and     denoting column breaks
         '''
-        history_str='APP_PARAMS\tCLI_COMMAND\tAPPLICATION\tMESSAGE\tOBJECT_ID\tOBSERVATION_ID\tORIGIN\tPRIORITY\tTIME\n'
+        history_str='APP_PARAMS;CLI_COMMAND;APPLICATION;MESSAGE;OBJECT_ID;OBSERVATION_ID;ORIGIN;PRIORITY;TIME\n'
         app_params=history_table.getcol('APP_PARAMS')['array']
         cli_command=history_table.getcol('CLI_COMMAND')['array']
         application=history_table.getcol('APPLICATION')
@@ -36,14 +39,14 @@ class MS(UVData):
         #Now loop through columns and generate history string
         for tbrow in range(len(times)):
             newline=str(app_params[tbrow]) \
-            +'\t'+str(cli_command[tbrow]) \
-            +'\t'+str(application[tbrow]) \
-            +'\t'+str(message[tbrow]) \
-            +'\t'+str(obj_id[tbrow]) \
-            +'\t'+str(obs_id[tbrow]) \
-            +'\t'+str(origin[tbrow]) \
-            +'\t'+str(priority[tbrow]) \
-            +'\t'+str(times[tbrow])+'\n'
+            +';'+str(cli_command[tbrow]) \
+            +';'+str(application[tbrow]) \
+            +';'+str(message[tbrow]) \
+            +';'+str(obj_id[tbrow]) \
+            +';'+str(obs_id[tbrow]) \
+            +';'+str(origin[tbrow]) \
+            +';'+str(priority[tbrow]) \
+            +';'+str(times[tbrow])+'\n'
             history_str+=newline
         return history_str
 
@@ -57,104 +60,131 @@ class MS(UVData):
 
 
 
-    def read_ms(self,filename,run_check=True,run_sanity_check=True,data_column='DATA'):
-    '''
-    read in a casa measurement set
-    ARGS:
+    def read_ms(self,filepath,run_check=True,run_sanity_check=True,data_column='DATA'):
+        '''
+        read in a casa measurement set
+        ARGS:
         filename: name of the measurement set folder
         run_check:
-    '''
-            if not os.path.exists(filepath):
-                raise(IOError, filepath + ' not found')
-            #set visibility units
-            if(data_column=='DATA'):
[…]
