### 2017-01-12 fc27f32
more bug fixes and test_ms.py
