### 2017-01-12 105d58d
more bug fixes and test_ms.py
