### 2017-01-12 c6510c4
more bug fixes and test_ms.py
