### 2017-01-12 b75fc6e
more bug fixes and test_ms.py
