### 2017-01-12 dd0fd85
more bug fixes and test_ms.py
