+++ uvdata/tests/test_ms.py
@@ -59,6 +59,7 @@ def test_ReadMSWriteReadUVFits(msname):
     #print uvfits_uv.vis_units
     #print uvfits_uv_orig.vis_units
     ms_uv.order_pols()
+
     nt.assert_equal(ms_uv, uvfits_uv)
     nt.assert_equal(uvfits_uv,uvfits_uv_orig)
     del(ms_uv)
