### 2017-01-12 c9ddea8
Add community guidelines (issue #122)
