+++ docs/index.rst
@@ -53,8 +53,19 @@ Known Issues and Planned Improvements
 -  support for direct reading and writing of Measurement Sets
 
 For details see the `issue
-log <https://github.com/HERA-Team/pyuvdata/issues>`__. Contributions are
-welcome!
+log <https://github.com/HERA-Team/pyuvdata/issues>`__.
+
+Community Guidelines
+--------------------
+
+Contributions are to this package to add new file formats or address any
+of the issues in the `issue
+log <https://github.com/HERA-Team/pyuvdata/issues>`__ are very welcome.
+Please submit improvements as pull requests against the repo.
+
+Bug reports or feature requests are also very welcome, please add them
+to the issue log after verifying that the issue does not already exist.
+Comments on existing issues are also welcome.
 
 History
 =======
