### 2017-01-12 HERA-Team/pyuvdata#122 @bhazelton
Addressed in latest commit.
