### 2017-01-12 14d4d1c
Cleaned up and expanded tutorial
