### 2017-01-12 b2ac3c6
tweak for .travis.yml
