+++ .travis.yml
@@ -3,7 +3,6 @@ python:
   - "2.7"
 # command to install dependencies
 install:
-  - pip install scipy
   - pip install .
 # command to run tests
 script: nosetests
