### 2017-01-13 d2c006f
more travis tweaks
