+++ .travis.yml
@@ -3,6 +3,7 @@ python:
   - "2.7"
 # command to install dependencies
 install:
+  - pip install --only-binary=numpy,scipy numpy scipy
   - pip install --user .
 # command to run tests
 script: nosetests
