+++ .travis.yml
@@ -3,6 +3,7 @@ python:
   - "2.7"
 # command to install dependencies
 install:
+  - pip install --upgrade pip setuptools wheel.
   - pip install --only-binary=numpy,scipy numpy scipy
   - pip install --user .
 # command to run tests
