### 2017-01-13 97e1826
another travis bug fix
