+++ .travis.yml
@@ -5,6 +5,6 @@ python:
 install:
   - pip install --upgrade pip setuptools wheel
   - pip install --only-binary=numpy,scipy numpy scipy
-  - pip install --user .
+  - pip install .
 # command to run tests
 script: nosetests
