### 2017-01-13 da023f8
add read_fits and write_fits functions into UVCal.
update nants_data_calculation
