### 2017-01-13 a79586d
add in parameters to read_calfits.
Add in switch for delay vs. gain data.
run_check after reading calibration file.
