### 2017-01-13 eddbdda
travis tweaks
