+++ .travis.yml
@@ -5,6 +5,8 @@ python:
 install:
   - pip install --upgrade pip setuptools wheel
   - pip install --only-binary=numpy,scipy numpy scipy
-  - pip install .
+  - pip install astropy
+  - pip install aipy
+  - python setup.py install
 # command to run tests
 script: nosetests
