### 2017-01-13 bfd14ea
Added order_pols() method to uvdata.py
