### 2017-01-13 63a23fa
Added order_pols() method to uvdata.py
