### 2017-01-13 e2d7b43
try travis with conda and pkgw's aipy conda package
