### 2017-01-13 7e3cab2
travis conda bug fix
