+++ .travis.yml
@@ -23,12 +23,13 @@ install:
 
   # create environment and install dependencies
   - conda create -q -n test-environment python=$TRAVIS_PYTHON_VERSION numpy scipy astropy nose pip
+
   - source activate test-environment
   - conda install -c pkgw-forge aipy
   - pip install coveralls
   - python setup.py install
-
 script: nosetests pyuvdata --with-coverage --cover-package=pyuvdata
 
 after_success:
     - coveralls
+
