### 2017-01-13 b88f811
travis conda bug fix
