+++ .travis.yml
@@ -1,9 +1,9 @@
 language: python
 python:
+  # We don't actually use the Travis Python, but this keeps it organized.
   - "2.7"
-# command to install dependencies
 install:
- - sudo apt-get update
+  - sudo apt-get update
   # We do this conditionally because it saves us some downloading if the
   # version is the same.
   - if [[ "$TRAVIS_PYTHON_VERSION" == "2.7" ]]; then
@@ -19,11 +19,10 @@ install:
   # Useful for debugging any issues with conda
   - conda info -a
 
-  # Replace dep1 dep2 ... with your dependencies
+  # create environment and install dependencies
   - conda create -q -n test-environment python=$TRAVIS_PYTHON_VERSION numpy scipy astropy
   - source activate test-environment
   - conda install -c pkgw aipy
   - python setup.py install
 
-# command to run tests
 script: nosetests
