### 2017-01-14 b20e533
travis conda fix
