+++ .travis.yml
@@ -22,6 +22,7 @@ install:
   # create environment and install dependencies
   - conda create -q -n test-environment python=$TRAVIS_PYTHON_VERSION numpy scipy astropy
   - source activate test-environment
+  - conda install astropy
   - conda install -c pkgw aipy
   - python setup.py install
 
