### 2017-01-14 59bf132
travis conda bug fix attempt
