+++ .travis.yml
@@ -20,7 +20,7 @@ install:
   - conda info -a
 
   # create environment and install dependencies
-  - conda create -q -n test-environment python=$TRAVIS_PYTHON_VERSION numpy scipy astropy
+  - conda create -q -n test-environment python=$TRAVIS_PYTHON_VERSION numpy scipy astropy nose
   - source activate test-environment
   - conda install astropy
   - conda install -c pkgw aipy
