### 2017-01-14 8bb6bca
travis conda bug fix attempt
