+++ .travis.yml
@@ -24,6 +24,7 @@ install:
   # create environment and install dependencies
   - conda create -q -n test-environment python=$TRAVIS_PYTHON_VERSION numpy scipy astropy nose pip
 
+
   - source activate test-environment
   - conda install -c pkgw-forge aipy
   - pip install coveralls
