### 2017-01-14 1ba953b
remove travis conda cruft
