### 2017-01-14 8332e7c
Add travis status badge to readme
