### 2017-01-14 b597e03
update docs, make travis badge display properly
