### 2017-01-14 8add55d
update docs, make travis badge display properly
