+++ docs/make_index.py
@@ -11,6 +11,11 @@ out = ('.. pyuvdata documentation master file, created by\n'
        '   make_index.py on {date}\n\n').format(date=t.iso)
 
 readme_text = pypandoc.convert_file('../readme.md', 'rst')
+
+rst_status_badge = '.. image:: https://travis-ci.org/HERA-Team/pyuvdata.svg?branch=master\n    :target: https://travis-ci.org/HERA-Team/pyuvdata'
+status_badge_text = '|Build Status|'
+readme_text = readme_text.replace(status_badge_text, rst_status_badge)
+
 end_text = 'parameters description'
 regex = re.compile(end_text.replace(' ', '\s+'))
 loc = re.search(regex, readme_text).start()
