### 2017-01-15 ff385f5
ms.py now treates Nant_telescope as the number of entries in the antenna positions table that have values not set equal to zero (rather than the total length of the antenna positions list)
