+++ uvdata/tests/test_ms.py
@@ -41,9 +41,9 @@ def test_ReadMSWriteReadUVFits(msname):
     #print uvfits_uv_orig.instrument
     #print uvfits_uv.object_name
     #print uvfits_uv_orig.object_name
-
     #print uvfits_uv.lst_array
     #print uvfits_uv_orig.lst_array
+
 
     #print uvfits_uv.phase_center_epoch
     #print uvfits_uv_orig.phase_center_epoch
