### 2017-01-18 494eeb9
Added all CTYPE headers for each image axis.
