### 2017-01-18 d8134de
added a test_calfits.py to test calfits module.
update test_cal.py tests.
