### 2017-01-20 2c7717f
update header parameters in writing calfits file.
Update calfits read function.

Update NAXISn headers. Remove Column's of tables.
Change axis of concatenation to the last axis.
