+++ uvdata/calfits.py
@@ -56,11 +56,7 @@ class CALFITS(UVCal):
         prihdr['SIMPLE'] = True
         prihdr['BITPIX'] = 32
         prihdr['NAXIS'] = 5
-        prihdr['NAXIS1'] = (self.Nants_data, 'Number and antennas')
-
-        prihdr['NAXIS3'] = (self.Ntimes, 'Number of time samples')
-        prihdr['NAXIS4'] = (self.Npols, 'Number of polarizations')
-
+        #prihdr['NAXIS1'] = (self.Nants_data, 'Number and antennas')
         prihdr['TELESCOP'] = ('HERA', 'Telescope of calibration')
         prihdr['OBSERVER'] = 'Observer'
         prihdr['DATE'] = today
@@ -77,89 +73,66 @@ class CALFITS(UVCal):
         prihdr['HISTORY'] = self.history
         prihdr['NSPWS'] = self.Nspws
         prihdr['XORIENT'] = self.x_orientation
-        prihdr['END']
 
         if np.all(self.gain_array):
-            coldat = fits.Column(name='GAIN', format='M',
-                                 array=self.gain_array)
-            colflg = fits.Column(name='FLAG', format='L',
-                                 array=self.flag_array)
-            colqual = fits.Column(name='QUALITY', format='D',
-                                  array=self.quality_array)
             # Set header variable for gain.
-            prihdr['NAXIS2'] = (self.Nfreqs, 'Number of frequency channels')
-            prihdr['CTYPE2'] = ('FREQS', 'Frequency.')
-            prihdr['CUNIT2'] = ('GHz', 'Unit of frequecy.')
-            prihdr['CRVAL2'] = self.freq_array[0]
-            prihdr['CDELT2'] = self.freq_array[1] - self.freq_array[0]
-            # set the last axis for number of arrays.
-            prihdr['NAXIS5'] = (4, 'Number of data arrays:gain.real, \
-                                    gain.imag, flag, quality')
-            prihdr['CTYPE5'] = ('Narrays', 'Number of image arrays.')
-            prihdr['CUNIT5'] = ('Integer', 'Number of image arrays. Increment.')
-            prihdr['CRVAL5'] = 0
-            prihdr['CDELT5'] = 1
-        elif np.all(self.delay_array):
-            coldat = fits.Column(name='DELAY', format='D',
-                                 array=self.delay_array)
-            colflg = fits.Column(name='FLAG', format='L',
-                                 array=self.flag_array)
-            colqual = fits.Column(name='QUALITY', format='D',
-                                  array=self.quality_array)
-            # Set header variable for gain.
-            prihdr['NAXIS2'] = (1, 'Number of delay solutions.')
-            prihdr['CTYPE2'] = ('DELAYS', 'Delay number.')
-            prihdr['CUNIT2'] = ('ns', 'Nanosecond units.')
-            prihdr['CRVAL2'] = 0
-            prihdr['CDELT2'] = 0
-            # set the last axis for number of arrays.
-            prihdr['NAXIS5'] = (3, 'Number of data arrays:delay,flag, quality')
-            prihdr['CTYPE5'] = ('Narrays', 'Number of image arrays.')
-            prihdr['CUNIT5'] = ('Integer', 'Number of image arrays. Increment.')
-            prihdr['CRVAL5'] = 0
-            prihdr['CDELT5'] = 1
+            prihdr['CTYPE4'] = ('FREQS', 'Frequency.')
+            prihdr['CUNIT4'] = ('GHz', 'Units of frequecy.')
+            prihdr['CRVAL4'] = self.freq_array[0][0]
+            prihdr['CDELT4'] = self.freq_array[0][1] - self.freq_array[0][0]
 
[…]
