### 2017-01-20 #140 @bhazelton
Thank you for the reminder. We are working on addressing the issues (we've
closed several of them), but we are a distributed team of scientists with
fairly busy research schedules so it may take a little bit longer. We
really appreciate the issues that were raised and we think they will make
the project better, so we want to address them fully and properly.

On Fri, Jan 20, 2017 at 9:38 AM, Arfon Smith <notifications@github.com>
wrote:

> Friendly reminder on this @bhazelton <https://github.com/bhazelton>.
>
