### 2017-01-24 5b6e710
First attempt at appveyor integration
