### 2017-01-24 ad193e5
appveyor bug fix
