+++ appveyor.yml
@@ -40,8 +40,8 @@ init:
 install:
   # this installs the appropriate Miniconda (Py2/Py3, 32/64 bit)
   # updates conda & installs: conda-build jinja2 anaconda-client
-  - powershell .\ci\install.ps1
-  - SET PATH=%PYTHON%;%PYTHON%\Scripts;%PATH%
+  - powershell .\\continuous-integration\\appveyor\\install.ps1
+  - "SET PATH=%PYTHON%;%PYTHON%\\Scripts;%PATH%"
   - echo "install"
   - cd
   - ls -ltr
