### 2017-01-24 d8a8388
Add required scripts for appveyor
