### 2017-01-24 5ffcf18
appveyor bug fix
