+++ appveyor.yml
@@ -19,7 +19,7 @@ environment:
 
   matrix:
 
-    - PYTHON: "C:\\Python27_64"
+    - CONDA_ROOT: "C:\\Miniconda3.5_64"
       PYTHON_VERSION: "2.7"
       PYTHON_ARCH: "64"
       CONDA_PY: "27"
