### 2017-01-24 e87613e
appveyor bug fix
