+++ appveyor.yml
@@ -52,7 +52,7 @@ install:
   - cmd: rmdir C:\cygwin /s /q
 
   # install our build environment
-  - cmd: %CONDA_BLD%\Scripts\conda config --set show_channel_urls true --set always_yes true --set changeps1 false
+  - "%CONDA_BLD%\\Scripts\\conda config --set show_channel_urls true --set always_yes true --set changeps1 false"
   - cmd: conda update -q conda
   # Useful for debugging any issues with conda
   - cmd: conda info -a
