### 2017-01-24 ca6eb70
appveyor -- more documentation, still doesn't work
