### 2017-01-26 6b933ab
Added table to store antenna names and antenna index's
