### 2017-01-26 c9686a8
updated name of file in test_cal.py
Tests are passing.
