+++ uvdata/tests/test_cal.py
@@ -85,7 +85,7 @@ class TestUVCalBasicMethods(object):
     def setUp(self):
         """Set up test"""
         self.uv_cal_object = UVCal()
-        self.testfile = os.path.join(DATA_PATH, 'zen.2457698.50098.xx.fits')
+        self.testfile = os.path.join(DATA_PATH, 'test123.fits')
         self.uv_cal_object.read_calfits(self.testfile)
         uvtest.checkWarnings(self.uv_cal_object.read_calfits, [self.testfile],
                              message='Telescope EVLA is not')
