### 2017-01-26 6e0d791
updated name of file in test_calfits.py.
Changed default number of warnings to 0.
Test is passing.
