### 2017-01-30 291122f
another appveyor attempt
