+++ appveyor.yml
@@ -10,6 +10,9 @@
 
 # Backslashes in quotes need to be escaped: \ -> "\\"
 
+matrix:
+  fast_finish: true     # immediately finish build once one of the jobs fails.
+
 environment:
   global:
     # SDK v7.0 MSVC Express 2008's SetEnv.cmd script will fail if the
@@ -18,7 +21,6 @@ environment:
     CMD_IN_ENV: "cmd /E:ON /V:ON /C .\\ci\\run_with_env.cmd"
 
   matrix:
-    fast_finish: true     # immediately finish build once one of the jobs fails.
 
     - PYTHON: "C:\\Python27_64"
       PYTHON_VERSION: "2.7"
