### 2017-01-31 d5194e1
change shape of delay array to be compatible with gain array.
Added to delay switch statements.
