### 2017-01-31 9c130e5
added some more tests for uvcal.
