### 2017-02-01 376d876
Adding more data conversions to tutorial. (Issue #119)
