### 2017-02-01 7adca30
Adding more data conversions to tutorial. (Issue #119)
