### 2017-02-01 89e3f17
Adding more data conversions to tutorial. (Issue #119)
