### 2017-02-01 cc15d11
appveyor -- try to install aipy via pip instead
