+++ appveyor.yml
@@ -65,7 +65,8 @@ install:
   # create our env
   - cmd: conda create -q -n test-environment python=%PYTHON_VERSION% astropy nose
   - cmd: activate test-environment
-  - cmd: conda install -c pkgw aipy
+  - cmd: pip install aipy
+#  - cmd: conda install -c pkgw aipy
   - cmd: python setup.py install
 
 test_script:
