### 2017-02-01 HERA-Team/pyuvdata#119 @adampbeardsley
The tutorial has been updated to the newer syntax, and expanded to include more file conversion examples and other functions, as well as plotting.
