### 2017-02-02 c0048bb
appveyor try cloning hera team aipy repo and installing directly
