### 2017-02-02 8464cbe
fix appveyor aipy download
