### 2017-02-02 9617ec1
appveyor.yml bug fix
