+++ appveyor.yml
@@ -67,7 +67,7 @@ install:
   - cmd: activate test-environment
 #  - cmd: pip install aipy
 #  - cmd: conda install -c pkgw aipy
-  - cmd: git clone -q --branch=master https://github.com:HERA-Team/aipy.git C:\projects\aipy
+  - cmd: git clone -q --branch=master https://github.com/HERA-Team/aipy.git C:\projects\aipy
   - cmd: cd ../aipy
   - cmd: python setup.py install
   - cmd: cd ../pyuvdata
