### 2017-02-06 d2c5173
Comment out lst array parameter for now
