### 2017-02-06 f9b1cc0
Add in try except statements for adding in frequency and time
parameters when writing calfits files.
