### 2017-02-08 3ca90d4
try setting distutils_debug environement variable for appveyor
