### 2017-02-08 5123b00
update test_calfits to clobber fits file if it exists.
Added clobber options to write_calfits functions.
