### 2017-02-08 a8d5099
tweak to distutils_debug variable for appveyor
