### 2017-02-08 38f49b7
attempt to switch to 32bit to avoid python 2.7 compilation error
