### 2017-02-08 1c2d1b4
actually switch to 32bit for appveyor
