+++ appveyor.yml
@@ -22,11 +22,11 @@ environment:
 
   matrix:
 
-    # don't use 64 bit on windows because of python compiling error.
+    # don't use 64 bit on windows because of python 2.7 compiling error.
     # see https://github.com/numpy/numpy/issues/6882
     - PYTHON: "C:\\Python27"
       PYTHON_VERSION: "2.7"
-      PYTHON_ARCH: "x86"
+      PYTHON_ARCH: "32"
       CONDA_PY: "27"
       CONDA_NPY: "110"
 
