### 2017-02-08 de0f0f8
still trying to get 32bit working for appveyor
