+++ appveyor.yml
@@ -31,11 +31,6 @@ environment:
       CONDA_NPY: "110"
 
 
-# We always use a 64-bit machine, but can build x86 distributions
-# with the PYTHON_ARCH variable (which is used by CMD_IN_ENV).
-platform:
-    - x64
-
 # all our python builds have to happen in tests_script...
 build: false
 
