### 2017-02-08 0df4a0a
try renaming python for conda in appveyor
