+++ appveyor.yml
@@ -24,7 +24,7 @@ environment:
 
     # don't use 64 bit on windows because of python 2.7 compiling error.
     # see https://github.com/numpy/numpy/issues/6882
-    - PYTHON: "C:\\Python27"
+    - PYTHON: "C:\\Python27_32"
       PYTHON_VERSION: "2.7"
       PYTHON_ARCH: "32"
       CONDA_PY: "27"
