### 2017-02-10 4672c7d
Use my fork of aipy for appveyor with platform independent paths
