### 2017-02-10 83feb82
Add some debugging info to appveyor
