### 2017-02-14 2994001
fixed developer docs by removing relative imports
