### 2017-02-14 5260d45
Add UVData class to developer docs
