+++ docs/developer_docs.rst
@@ -12,6 +12,9 @@ Classes
 .. autoclass:: uvbase.UVBase
   :members:
 
+.. autoclass:: uvdata.UVData
+  :members:
+
 .. autoclass:: uvfits.UVFITS
   :members:
 
