### 2017-02-14 907202b
Add UVData class to developer docs
