### 2017-02-14 7c768de
slight readme tweaks
