### 2017-02-14 87f26c5
slight readme tweaks
