### 2017-02-14 79d906e
docs cleanup, tweaks
