### 2017-02-14 2715047
docs cleanup, tweaks
