### 2017-02-14 ef2aeaa
docs cleanup, tweaks
