+++ pyuvdata/miriad.py
@@ -16,6 +16,7 @@ class Miriad(UVData):
     and write_miriad methods on the UVData class.
     """
 
+
     def _pol_to_ind(self, pol):
         if self.polarization_array is None:
             raise(ValueError, "Can't index polarization {p} because "
