+++ docs/index.rst
@@ -161,4 +161,5 @@ Further Documentation
    uvcal_parameters
 
    uvcal
+
    developer_docs
