### 2017-02-14 HERA-Team/pyuvdata#120 @bhazelton
This is fixed in the latest commit.
