### 2017-02-14 84d4973
docs tweaks & improvements
