### 2017-02-14 5a881f7
docs tweaks & improvements
