### 2017-02-14 23638ba
rename main directory to pyuvdata
