### 2017-02-14 003d392
update tests and tutorials
