+++ pyuvdata/tests/__init__.py
@@ -5,7 +5,6 @@ import sys
 from pyuvdata.data import DATA_PATH
 import pyuvdata.utils as uvutils
 
-
 def setup_package():
     """Make data/test directory to put test output files in."""
     testdir = os.path.join(DATA_PATH, 'test/')
