### 2017-02-14 ff55379
update tests and tutorials
