+++ pyuvdata/tests/test_parameter.py
@@ -1,132 +1,133 @@
 import nose.tools as nt
 import numpy as np
-import uvdata
+from pyuvdata import parameter as uvp
+from pyuvdata import UVBase
 
 
 def test_class_inequality():
     """Test equality error for different uvparameter classes."""
-    param1 = uvdata.UVParameter(name='p1', value=1)
-    param2 = uvdata.AngleParameter(name='p2', value=1)
+    param1 = uvp.UVParameter(name='p1', value=1)
+    param2 = uvp.AngleParameter(name='p2', value=1)
     nt.assert_not_equal(param1, param2)
 
 
 def test_value_class_inequality():
     """Test equality error for different uvparameter classes."""
-    param1 = uvdata.UVParameter(name='p1', value=3)
-    param2 = uvdata.UVParameter(name='p2', value=np.array([3, 4, 5]))
+    param1 = uvp.UVParameter(name='p1', value=3)
+    param2 = uvp.UVParameter(name='p2', value=np.array([3, 4, 5]))
     nt.assert_not_equal(param1, param2)
     nt.assert_not_equal(param2, param1)
-    param3 = uvdata.UVParameter(name='p2', value='Alice')
+    param3 = uvp.UVParameter(name='p2', value='Alice')
     nt.assert_not_equal(param1, param3)
 
 
 def test_array_inequality():
     """Test equality error for different array values."""
-    param1 = uvdata.UVParameter(name='p1', value=np.array([0, 1, 3]))
-    param2 = uvdata.UVParameter(name='p2', value=np.array([0, 2, 4]))
+    param1 = uvp.UVParameter(name='p1', value=np.array([0, 1, 3]))
+    param2 = uvp.UVParameter(name='p2', value=np.array([0, 2, 4]))
     nt.assert_not_equal(param1, param2)
-    param3 = uvdata.UVParameter(name='p3', value=np.array([0, 1]))
+    param3 = uvp.UVParameter(name='p3', value=np.array([0, 1]))
     nt.assert_not_equal(param1, param3)
 
 
 def test_string_inequality():
     """Test equality error for different string values."""
-    param1 = uvdata.UVParameter(name='p1', value='Alice')
-    param2 = uvdata.UVParameter(name='p2', value='Bob')
+    param1 = uvp.UVParameter(name='p1', value='Alice')
+    param2 = uvp.UVParameter(name='p2', value='Bob')
     nt.assert_not_equal(param1, param2)
 
 
 def test_string_list_inequality():
     """Test equality error for different string values."""
-    param1 = uvdata.UVParameter(name='p1', value=['Alice', 'Eve'])
-    param2 = uvdata.UVParameter(name='p2', value=['Bob', 'Eve'])
+    param1 = uvp.UVParameter(name='p1', value=['Alice', 'Eve'])
+    param2 = uvp.UVParameter(name='p2', value=['Bob', 'Eve'])
     nt.assert_not_equal(param1, param2)
 
 
 def test_string_equality():
     """Test equality error for different string values."""
-    param1 = uvdata.UVParameter(name='p1', value='Alice')
-    param2 = uvdata.UVParameter(name='p2', value='Alice')
+    param1 = uvp.UVParameter(name='p1', value='Alice')
+    param2 = uvp.UVParameter(name='p2', value='Alice')
     nt.assert_equal(param1, param2)
 
 
 def test_integer_inequality():
     """Test equality error for different non-array, non-string values."""
-    param1 = uvdata.UVParameter(name='p1', value=1)
-    param2 = uvdata.UVParameter(name='p2', value=2)
+    param1 = uvp.UVParameter(name='p1', value=1)
[…]
