+++ setup.py
@@ -28,6 +28,7 @@ setup_args = {
                     'Topic :: Scientific/Engineering :: Astronomy'],
     'keywords': 'radio astronomy interferometry'
 
+
 }
 
 if __name__ == '__main__':
