### 2017-02-14 589260c
Restore pyephem as a dependency in the readme
Accidentally removed earlier
