### 2017-02-14 9ff0ce9
Restore pyephem as a dependency in the readme
Accidentally removed earlier
