+++ docs/index.rst
@@ -95,6 +95,7 @@ install.
 -  numpy >= 1.10
 -  scipy
 -  astropy >= 1.2
+-  pyephem
 -  aipy
 
 Install pyuvdata
