### 2017-02-14 dd2c0e4
Restore pyephem as a dependency in the readme
Accidentally removed earlier
