+++ scripts/test_coverage.sh
@@ -5,5 +5,5 @@ cd $DIR/..
 
 python setup.py install
 
-cd uvdata/tests
-nosetests --with-coverage --cover-erase --cover-package=uvdata --cover-html "$@"
+cd pyuvdata/tests
+nosetests --with-coverage --cover-erase --cover-package=pyuvdata --cover-html "$@"
