### 2017-02-15 f989216
move appveyor.yml out of main directory until aipy dependency is removed
