+++ appveyor.yml
@@ -1,81 +1,0 @@
-# based on PANDAS appveyor.yml
-# https://searchcode.com/file/116370094/appveyor.yml
-# Which references the following sites as inspiration
-#   http://tjelvarolsson.com/blog/how-to-continuously-test-your-python-code-on-windows-using-appveyor/
-#   https://packaging.python.org/en/latest/appveyor/
-#   https://github.com/rmcgibbo/python-appveyor-conda-example
-
-# Also looked at:
-# https://github.com/pythonnet/pythonnet/blob/master/appveyor.yml
-
-# Backslashes in quotes need to be escaped: \ -> "\\"
-
-matrix:
-  fast_finish: true     # immediately finish build once one of the jobs fails.
-
-environment:
-  global:
-    # SDK v7.0 MSVC Express 2008's SetEnv.cmd script will fail if the
-    # /E:ON and /V:ON options are not enabled in the batch script intepreter
-    # See: http://stackoverflow.com/a/13751649/163740
-    CMD_IN_ENV: "cmd /E:ON /V:ON /C .\\ci\\run_with_env.cmd"
-
-  matrix:
-
-    # don't use 64 bit on windows because of python 2.7 compiling error.
-    # see https://github.com/numpy/numpy/issues/6882
-    - PYTHON: "C:\\Python27_32"
-      PYTHON_VERSION: "2.7"
-      PYTHON_ARCH: "32"
-      CONDA_PY: "27"
-      CONDA_NPY: "110"
-
-
-# all our python builds have to happen in tests_script...
-build: false
-
-init:
-  - "ECHO %PYTHON_VERSION% %PYTHON%"
-
-install:
-  # this installs the appropriate Miniconda (Py2/Py3, 32/64 bit)
-  # updates conda & installs: conda-build jinja2 anaconda-client
-  - powershell .\ci\install.ps1
-  - SET PATH=%PYTHON%;%PYTHON%\Scripts;%PATH%
-  - echo "install"
-  - cd
-  - ls -ltr
-  - git tag --sort v:refname
-
-  # this can conflict with git
-  - cmd: rmdir C:\cygwin /s /q
-
-  # install our build environment
-  - cmd: conda config --set show_channel_urls true --set always_yes true --set changeps1 false
-  - cmd: conda update -q conda
-  # Useful for debugging any issues with conda
-  - cmd: conda info -a
-
-  # build em using the local source checkout in the correct windows env
-  - cmd: '%CMD_IN_ENV% conda build ci\appveyor.recipe -q'
-
-  # create our env
-  - cmd: conda create -q -n test-environment python=%PYTHON_VERSION% numpy scipy astropy nose
-  - cmd: activate test-environment
-#  - cmd: pip install aipy
-#  - cmd: conda install -c pkgw aipy
-  - cmd: set DISTUTILS_DEBUG='test'
-  - cmd: ls -ltr "C:\\Program Files (x86)\\Microsoft Visual Studio 9.0\\VC\\BIN\\"
-#  - cmd: git clone -q --branch=master https://github.com/HERA-Team/aipy.git C:\projects\aipy
-  - cmd: git clone -q --branch=master https://github.com/bhazelton/aipy.git C:\projects\aipy
-  - cmd: cd ../aipy
-  - cmd: python setup.py install
-  - cmd: cd ../pyuvdata
-  - cmd: python setup.py install
-
-test_script:
-  # tests
-  - cd \
-  - cmd: activate test-environment
-  - cmd: conda list
-  - cmd: nosetests
