### 2017-02-15 c3ac9c1
move appveyor.yml out of main directory until aipy dependency is removed
