+++ docs/uvdata.rst
@@ -4,5 +4,5 @@ UVData Class
 UVData is the main user class. It provides import and export functionality to
 all supported file formats (UVFITS, Miriad, FHD) and can be interacted with directly.
 
-.. autoclass:: uvdata.UVData
+.. autoclass:: pyuvdata.UVData
   :members:
