### 2017-02-15 bbf80d6
Removed language using "sane" or "sanity"
