+++ pyuvdata/telescopes.py
@@ -48,7 +48,7 @@ class Telescope(uvbase.UVBase):
                 'telescope_location_lat_lon_alt_degrees properties')
         self._telescope_location = uvp.LocationParameter('telescope_location',
                                                          description=desc,
-                                                         sane_range=(6.35e6, 6.39e6),
+                                                         acceptable_range=(6.35e6, 6.39e6),
                                                          tols=1e-3)
         # possibly add in future versions:
         # Antenna positions (but what about reconfigurable/growing telescopes?)
@@ -90,7 +90,7 @@ def get_telescope(telescope_name):
                                                   telescope_dict['longitude'],
                                                   telescope_dict['altitude'])
 
-            obj.check(run_sanity_check=True)
+            obj.check(run_check_acceptability=True)
     else:
         # no telescope matching this name
         return False
