### 2017-02-15 5b38ce1
Removed language using "sane" or "sanity"
