### 2017-02-15 99a9c0e
Removed language using "sane" or "sanity"
