+++ uvdata/__init__.py
@@ -1,0 +1,3 @@
+"""init file for uvdata."""
+
+raise ImportError('uvdata is depricated. Please use "import pyuvdata".')
