+++ setup.py
@@ -9,8 +9,8 @@ setup_args = {
     'name': 'pyuvdata',
     'author': 'HERA Team',
     'license': 'BSD',
-    'package_dir': {'pyuvdata': 'pyuvdata'},
-    'packages': ['pyuvdata'],
+    'package_dir': {'pyuvdata': 'pyuvdata', 'uvdata': 'uvdata'},
+    'packages': ['pyuvdata', 'uvdata'],
     'scripts': glob.glob('scripts/*'),
     'version': __version__,
     'package_data': {'pyuvdata': [f for f in listdir('./pyuvdata/data') if op.isfile(op.join('./pyuvdata/data', f))]},
