### 2017-02-15 b4b39dd
Adding deprication warning for uvdata
