### 2017-02-15 da3d5d4
Fixed test run for travis and in readme
