### 2017-02-15 e951328
Updated parameter doc to correct shape for uvw_array
