### 2017-02-15 538a0e1
config tweaks for pypi
