### 2017-02-15 efe8e73
config tweaks for pypi
