### 2017-02-15 c748da7
syntax bug fix in setup.py
