### 2017-02-15 b976e6f
syntax bug fix in setup.py
