### 2017-02-16 dbcab34
added a manifest, short description and version update for pypi
