### 2017-02-16 7df9f49
added a manifest, short description and version update for pypi
