### 2017-02-16 HERA-Team/pyuvdata#117 @bhazelton
This is now fixed. pyuvdata is on pypi and the readme doesn't have broken links
