### 2017-02-16 #140 @bhazelton
@arfon @astrofrog I think we have now fixed all the issues that were raised, both critical and non-critical.
