### 2017-02-16 #140 @astrofrog
This looks good to me now, so as far as I'm concerned this can be accepted!

cc @arfon @whedon
