### 2017-02-16 38f4df5
add in cal and calfits imports to pyuvdata/__init__.py
