### 2017-02-16 28c61d8
removed language using "sane" or "sanity" from cal.py and calfits.py
