### 2017-02-16 36f17ca
Add docs for cal stuff
