### 2017-02-16 fea4f09
Add docs for cal stuff
