+++ docs/cal_parameters.rst
@@ -42,6 +42,7 @@ are required for most kinds of uv cal files.
 **flag_array**
      Array of flags to be applied to calibrated data (logical OR                  of input and flag generated by calibration). True is flagged.shape: (Nants_data, Nfreqs, Ntimes, Npols), type = bool.
 
+
 **freq_array**
      Array of frequencies, shape (Nspws, Nfreqs), units Hz
 
@@ -58,10 +59,12 @@ are required for most kinds of uv cal files.
      Array of polarization integers, shape (Npols). AIPS Memo 117 says: stokes 1:4 (I,Q,U,V);  circular -1:-4 (RR,LL,RL,LR); linear -5:-8 (XX,YY,XY,YX)
 
 **quality_array**
+
      Array of qualities of calibration solutions                 shape: (Nants_data, Nfreqs, Ntimes, Npols), type = float.
 
 **telescope_name**
      Name of telescope. e.g. HERA. String.
+
 
 **time_array**
      Array of times, center of integration, shape (Ntimes), units Julian Date
