### 2017-02-17 df67d92
update make_index.py to be read in branch specific travis badge details
