### 2017-02-17 e0c6e4e
update make_index.py to be read in branch specific travis badge details
