+++ docs/make_index.py
@@ -8,6 +8,7 @@ import pypandoc
 from astropy.time import Time
 
 
+
 def write_index_rst(readme_file=None, write_file=None):
     t = Time.now()
     t.out_subfmt = 'date'
@@ -67,4 +68,5 @@ def write_index_rst(readme_file=None, write_file=None):
     F = open(write_file, 'w')
     F.write(out)
     print("wrote " + write_file)
+
 
