### 2017-02-20 #140 @bhazelton
@arfon Here's the Zenodo badge: [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.297348.svg)](https://doi.org/10.5281/zenodo.297348)
