### 2017-02-21 eb9c30e
Remove old cruft
