### 2017-02-21 bd9e81a
Fix for astropy 1.3 deprecation warning (issue #128)
needs to use 'clobber' for astropy 1.2 and 'overwrite' for astropy 1.3. Added a test for the version to determine what to use
