### 2017-02-22 b7ddbc4
Made changes to cal.py from code rev.
add channel_width, integration_time, git origin/hash parameters.
updated doc string of flags and quality array.
updated switch statements so flag parameters stay as shape of gain (w/ freq dependence).
