+++ pyuvdata/cal.py
@@ -61,6 +61,12 @@ class UVCal(UVBase):
                                            expected_type=np.float,
                                            tols=1e-3)  # mHz
 
+        desc = ('Channel width of of a frequency bin. Units Hz.')
+        self._channel_width = uvp.UVParameter('channel_width',
+                                              description=desc,
+                                              expected_type=np.float,
+                                              tols=1e-3)
+
         desc = ('Array of polarization integers, shape (Npols). '
                 'AIPS Memo 117 says: stokes 1:4 (I,Q,U,V);  '
                 'circular -1:-4 (RR,LL,RL,LR); linear -5:-8 (XX,YY,XY,YX)')
@@ -77,12 +83,10 @@ class UVCal(UVBase):
                                            expected_type=np.float,
                                            tols=1e-3 / (60.0 * 60.0 * 24.0))
 
-#        desc = ('Array of lsts, corresponding to the time array,' +
-#                ' shape is (Ntimes), units Julian Date')
-#        self._lst_array = uvp.UVParameter('lst_array', description=desc,
-#                                          form=('Ntimes',),
-#                                          expected_type=np.float,
-#                                          tols=radian_tol)
+        desc = ('Integration time of a time bin. Units in JD')
+        self.integration_time = uvp.UVParameter('integration time',
+                                                description=desc,
+                                                expected_type=np.float)
 
         desc = ('The convention for applying he calibration solutions to data.'
                 'Indicates that to calibrate one should divide or multiply '
@@ -92,19 +96,22 @@ class UVCal(UVBase):
                                                 description=desc,
                                                 acceptable_vals=['divide', 'multiply'])
 
-        desc = ('Array of flags, True is flagged.'
+        desc = ('Array of flags to be applied to calibrated data (logical OR \
+                 of input and flag generated by calibration). True is flagged.'
                 'shape: (Nants_data, Nfreqs, Ntimes, Npols), type = bool.')
         self._flag_array = uvp.UVParameter('flag_array', description=desc,
                                            form=('Nants_data', 'Nfreqs',
                                                  'Ntimes', 'Npols'),
                                            expected_type=np.bool)
 
-        desc = ('Array of qualities, shape: (Nants_data, Nfreqs, Ntimes, '
+        desc = ('Array of qualities of calibration solutions \
+                shape: (Nants_data, Nfreqs, Ntimes, '
                 'Npols), type = float.')
         self._quality_array = uvp.UVParameter('quality_array', description=desc,
                                               form=('Nants_data', 'Nfreqs',
                                                     'Ntimes', 'Npols'),
                                               expected_type=np.float)
+
         desc = ('Orientation of the physical dipole corresponding to what is '
                 'labelled as the x polarization. Values are east '
                 '(east/west orientation),  north (north/south orientation) or '
@@ -142,6 +149,20 @@ class UVCal(UVBase):
                                                  form=('Nants_data', 'Nfreqs',
                                                        'Ntimes', 'Npols'),
                                                  expected_type=np.bool)
+
+        desc = ('Origin (on github for e.g) of calibration software. ')
+        self._git_origin = uvp.UVParameter('git_origin', form='str',
+                                           expected_type=str, value='none',
+                                           description=desc,
+                                           required=False)
+
+        desc = ('Commit hash of calibration software (from git_origin) used
+                'to generate solutions.')
+        self._git_hash = uvp.UVParameter('git_hash', form='str',
+                                         expected_type=str, value='none',
+                                         description=desc,
+                                         required=False)
+
         super(UVCal, self).__init__()
 
[…]
