### 2017-02-22 #140 @arfon
@astrofrog many thanks for the review here ✨

@bhazelton - your paper is now accepted into JOSS and your DOI is http://dx.doi.org/10.21105/joss.00140 :zap: 🚀 💥
